move batch_iter to a separate module
